requirements.txt
python-dotenv
cohere
selenium
requests
groq
AppOpener
pywhatkit
bs4
pillow
rich
keyboard
mtranslate
pygame
edge-tts
PyQt5
pyautogui
webdriver-manager
googlesearch-python